refactor(ai-engine): resolve #1766 — split packaging agent.py + validator.py into focused modules
Decompose the packaging/ subpackage monoliths that regrew to ~30K each:

agent.py (30K, 783 lines) -> thin 2.6K shim delegating to:
  - orchestrator.py (271 lines): PackagingAgent coordinator + state
  - assembler.py (279 lines): execution logic (manifests, blocks/items,
    entities, packaging, validation entry points)
  - tools.py (363 lines): Pydantic args_schema + typed BaseTool wrappers

validator.py (29K, 794 lines) -> thin 11.4K orchestrator composing:
  - schema_validator.py (448 lines): ValidationSeverity/Issue/Result data
    classes + SchemaValidatorMixin (structure/manifest/component checks)
  - runtime_validator.py (93 lines): RuntimeValidatorMixin (compatibility,
    forbidden files, version comparison)
  - validator.py (329 lines): PackagingValidator(SchemaValidatorMixin,
    RuntimeValidatorMixin) orchestration + scoring + reporting

Backwards compatibility preserved: agent.py and validator.py remain as
re-export shims, so all existing imports (agents.packaging_agent,
agents.packaging_validator, agents.packaging.*) work unchanged.

Tests: 895 passed, 1 unrelated pre-existing failure, 29 skipped (identical
to baseline). Lint clean on all new/modified files.
